+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0 - 2020-12-15
+
+### Added
+- Configurable filters
+- Configurable filter callback
+
 ## 0.2.0 - 2020-11-17
 
 ### Changed